SemanticTokens: Use size_t as argument for type
Should fix the case of idx = -1 being rightfully out of bounds.

BUG: 447553
(cherry picked from commit 6193824fbb2fd0d4a3ae829ee258c9d3268829cf)

+++ addons/lspclient/semantic_tokens_legend.h
@@ -51,7 +51,7 @@ public:
      */
     void initialize(const std::vector<QString> &types);
 
-    KTextEditor::Attribute::Ptr attributeForTokenType(int idx) const
+    KTextEditor::Attribute::Ptr attributeForTokenType(size_t idx) const
     {
         if (idx >= totalTokenTypes) {
             return {};
@@ -68,7 +68,7 @@ private:
     Q_SLOT void themeChange(KTextEditor::Editor *e);
     void refresh(const std::vector<TokenType> &m_tokenTypes);
 
-    int totalTokenTypes;
+    size_t totalTokenTypes;
     std::vector<KTextEditor::Attribute::Ptr> sharedAttrs;
     KTextEditor::Attribute::Ptr fixedAttrs[7];
 };
